requirements.txt
requests~=2.32.3
beautifulsoup4~=4.12.3
pandas~=2.2.2
psycopg2-binary~=2.9.9
sqlalchemy~=2.0.30
google-auth~=2.29.0
gspread~=6.0.0
python-dotenv~=1.0.0
pytest~=8.2.2
pytest-cov~=5.0.0
requests-mock~=1.12.1
python-dotenv>=0.21.0
pytest-mock==3.14.0